- Home: Add Best Seller Screen and View Model
This commit introduces the Best Seller screen and its corresponding view model. It includes changes to the home screen to navigate to the Best Seller screen, as well as updates to the web services and dependency injection configurations.
