+++ lib/di/injetible_intinalizer.config.dart
@@ -18,47 +18,37 @@ import '../data/data_source/remote_data_source/auth_remote_data_source/auth_remo
     as _i658;
 import '../data/data_source/remote_data_source/auth_remote_data_source/auth_remote_data_source_implementation.dart'
     as _i793;
-
 import '../data/data_source/remote_data_source/home_remote_data_source/best_seller_contract.dart'
     as _i694;
 import '../data/data_source/remote_data_source/home_remote_data_source/best_seller_impl.dart'
     as _i463;
-import '../data/repository_implementation/auth_repository_implementation/auth_repository_implementation.dart'
-    as _i277;
-import '../data/repository_implementation/home_repository_implementation/best_seller_rep_impl.dart'
-    as _i384;
-import '../data/web_services/WebServices.dart' as _i995;
-import '../domain/repository_contract/auth_repository_contract/auth_repository_contract.dart'
-    as _i284;
-import '../domain/repository_contract/home_repository_contract/best_seller_repository.dart'
-    as _i670;
-
 import '../data/data_source/remote_data_source/home_remote_data_source/home_remote_data_source_contract.dart'
     as _i680;
 import '../data/data_source/remote_data_source/home_remote_data_source/home_remote_data_source_implementation.dart'
     as _i782;
 import '../data/repository_implementation/auth_repository_implementation/auth_repository_implementation.dart'
     as _i277;
+import '../data/repository_implementation/home_repository_implementation/best_seller_rep_impl.dart'
+    as _i384;
 import '../data/repository_implementation/home_repository_implementation/home_repository_implementation.dart'
     as _i645;
 import '../data/web_services/WebServices.dart' as _i995;
 import '../domain/repository_contract/auth_repository_contract/auth_repository_contract.dart'
     as _i284;
+import '../domain/repository_contract/home_repository_contract/best_seller_repository.dart'
+    as _i670;
 import '../domain/repository_contract/home_repository_contract/home_repository_contract.dart'
     as _i1053;
-
 import '../domain/use_case/auth_use_case/forget_password_use_case.dart'
     as _i439;
 import '../domain/use_case/auth_use_case/login_use_case.dart' as _i6;
 import '../domain/use_case/auth_use_case/reset_password_use_case.dart' as _i455;
 import '../domain/use_case/auth_use_case/signup_use_case.dart' as _i179;
 import '../domain/use_case/auth_use_case/verify_reset_use_case.dart' as _i86;
-
 import '../domain/use_case/home_use_case/best_seller_use_case.dart' as _i554;
-
 import '../domain/use_case/home_use_case/home_use_case.dart' as _i817;
-
 import '../presentation/auth/cubit/auth_view_model.dart' as _i851;
+import '../presentation/home/tabs/home/best_seller_view_model.dart' as _i891;
 import '../presentation/home/tabs/home/home_view_model.dart' as _i540;
 
 extension GetItInjectableX on _i174.GetIt {
@@ -82,14 +72,13 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i658.AuthRemoteDataSourceContract>(
       () => _i793.AuthRemoteDataSourceImplementation(gh<_i995.WebServices>()),
     );
+    gh.factory<_i680.HomeRemoteDataSourceContract>(
+      () => _i782.HomeRemoteDataSourceImplementation(gh<_i995.WebServices>()),
+    );
     gh.factory<_i670.BestSellerRepository>(
       () => _i384.bestSellerRepositoryImpl(
         bestSellerContractt: gh<_i694.BestSellerContract>(),
       ),
-    );
-
-    gh.factory<_i680.HomeRemoteDataSourceContract>(
-      () => _i782.HomeRemoteDataSourceImplementation(gh<_i995.WebServices>()),
     );
     gh.factory<_i1053.HomeRepositoryContract>(
       () => _i645.HomeRepositoryImplementation(
@@ -99,7 +88,6 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i817.HomeUseCase>(
       () => _i817.HomeUseCase(gh<_i1053.HomeRepositoryContract>()),
     );
-
     gh.factory<_i284.AuthRepositoryContract>(
       () => _i277.AuthRepositoryImplementation(
         gh<_i658.AuthRemoteDataSourceContract>(),
@@ -130,8 +118,9 @@ extension GetItInjectableX on _i174.GetIt {
         obj_login_repository_contract: gh<_i284.AuthRepositoryContract>(),
       ),
     );
-    gh.singleton<_i540.HomeViewModel>(
-      () => _i540.HomeViewModel(BestSeller: gh<_i554.bestSellerUseCase>()),
+    gh.singleton<_i891.BestSellerViewModel>(
+      () =>
+          _i891.BestSellerViewModel(BestSeller: gh<_i554.bestSellerUseCase>()),
     );
     gh.singleton<_i851.AuthViewModel>(
       () => _i851.AuthViewModel(
