+++ lib/presentation/home/tabs/home/best_seller_states.dart
@@ -1,0 +1,14 @@
+
+import 'package:flowery/domain/entity/home_entity/best_seller_entity/best_seller_entity.dart';
+
+abstract class BestSellerState{
+}
+
+class Initial extends BestSellerState{}
+
+class IsLoadingBestSeller extends BestSellerState{}
+class SuccessBestSeller extends BestSellerState{
+  BestSellerEntity bestSellerEntity;
+  SuccessBestSeller({required this.bestSellerEntity});
+}
+class ErrorBestSeller extends BestSellerState{}
