+++ lib/presentation/home/tabs/home/home_states.dart
@@ -1,18 +1,4 @@
 
-import 'package:flowery/domain/entity/home_entity/best_seller_entity/best_seller_entity.dart';
-
-abstract class HomeStates{
-}
-
-class Initial extends HomeStates{}
-
-class IsLoadingBestSeller extends HomeStates{}
-class SuccessBestSeller extends HomeStates{
-  BestSellerEntity bestSellerEntity;
-  SuccessBestSeller({required this.bestSellerEntity});
-}
-class ErrorBestSeller extends HomeStates{}
-
 import 'package:equatable/equatable.dart';
 import 'package:flowery/data/model/home_model/categories/categories_model.dart';
 import 'package:flowery/data/model/home_model/home_model.dart';
